% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.gcproc.R
\name{cv.gcproc}
\alias{cv.gcproc}
\title{Generalised Canonical Procrustes}
\usage{
cv.gcproc(
  x,
  y,
  config = list(k_dim = 70, j_dim = 70, eta = 0.005, max_iter = 1500, min_iter = 15,
    tol = 0.001, log = F, center = T, scale.z = T, batches = 16, cores = 2, verbose = T,
    init = "svd"),
  anchors = NULL,
  pivots = NULL,
  initial_starts = 3
)
}
\arguments{
\item{x}{Reference dataset of cell by gene matrix}

\item{y}{Experimental dataset of cell by gene matrix}

\item{config}{Configuration parameters (please read gcproc code for more details)}

\item{anchors}{Fixing and anchoring the main model parameters to transfer prior information}

\item{pivots}{Initialisation of the main model parameters to speed up learning process}

\item{initial_starts}{Multiple fixed seeds for initial seeding}
}
\value{
The best seeded gcproc model with the highest negative log-likelihood is returned.
}
\description{
Runs gcproc with multiple initial seeds
}
