% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_config.R
\name{extract_anchors_framework}
\alias{extract_anchors_framework}
\title{Extract anchor framework to put into gcproc}
\usage{
extract_anchors_framework(verbose = T)
}
\arguments{
\item{anchor_y.sample}{Equivalent to alpha.L.K}

\item{anchor_y.feature}{Equivalent to v.beta}

\item{anchor_x.sample}{Equivalent to alpha.L.J}

\item{anchor_x.feature}{Equivalent to u.beta}

\item{anchor_y.cov.sample}{Equivalent to y.gamma}

\item{anchor_y.cov.feature}{Equivalent to y.delta}

\item{anchor_x.cov.sample}{Equivalent to x.gamma}

\item{anchor_x.cov.feature}{Equivalent to x.delta}
}
\value{
Anchor framework for gcproc
}
\description{
Anchors allow the transfer of learned parameters from a pre-trained model.
NOTE: This is an empty framework that the user must fill in.
}
