% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_config.R
\name{extract_config}
\alias{extract_config}
\title{Extract configuration parameters of gcproc}
\usage{
extract_config(verbose = T)
}
\arguments{
\item{verbose}{Print statements?}

\item{i_dim}{Dimension reduction for samples (assumed to be along rows)}

\item{j_dim}{Dimension reduction for features (assumed to be along columns)}

\item{min_iter}{Minimum iteration of gcproc}

\item{max_iter}{Maximum iteration of gcproc}

\item{n_cores}{Number of CPU cores, used for prediction only}

\item{tol}{Tolerance threshold for convergence (metric: Root Mean Squared Error)}

\item{init}{Initialisation method for the model ("random","eigen-quick","eigen-dense","svd-quick","svd-dense")}
}
\value{
Configuration parameters for gcproc
}
\description{
Extract configuration parameters of gcproc
}
