% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_config.R
\name{extract_covariate_framework}
\alias{extract_covariate_framework}
\title{Extract covariate framework to put into gcproc}
\usage{
extract_covariate_framework(verbose = T)
}
\arguments{
\item{factors}{Table of factors to put into gcproc - for each factor, there is a unique code}
}
\description{
Covariate data to improve modelling capacity for similar axes
}
