% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_config.R
\name{extract_covariates_framework}
\alias{extract_covariates_framework}
\title{Extract covariates framework to put into gcproc}
\usage{
extract_covariates_framework(verbose = T)
}
\arguments{
\item{covariates_y.sample}{Covariates with length equivalent to samples of y}

\item{covariates_y.feature}{Covariates with length equivalent to features of y}

\item{covariates_x.sample}{Covariates with length equivalent to samples of x}

\item{covariates_x.feature}{Covariates with length equivalent to features of x}
}
\value{
Covariate framework for gcproc
}
\description{
Covariates can be used to further improve batch modelling.
NOTE: This is an empty framework that the user must fill in.
}
