% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_config.R
\name{extract_join_framework}
\alias{extract_join_framework}
\title{Extract join framework to put into gcproc}
\usage{
extract_join_framework(verbose = T)
}
\arguments{
\item{alpha}{Joining the alpha parameters. A vector of integers, where identical integers indicate same the data axis to be joined. Axes that should not be shared are given NA.}

\item{beta}{Joining the beta parameters. A vector of integers, where identical integers indicate same the data axis to be joined. Axes that should not be shared are given NA.}
}
\description{
Join data to improve modelling capacity for similar axes
}
