% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_config.R
\name{extract_pivots_framework}
\alias{extract_pivots_framework}
\title{Extract pivot framework to put into gcproc.}
\usage{
extract_pivots_framework(verbose = T)
}
\arguments{
\item{code}{code}
}
\value{
Pivot framework for gcproc
}
\description{
Pivots allow initialisation of parameters as input.
NOTE: This is an empty framework that the user must fill in.
}
