% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_config.R
\name{extract_prediction_framework}
\alias{extract_prediction_framework}
\title{Extract prediction framework to put into gcproc}
\usage{
extract_prediction_framework(verbose = T)
}
\arguments{
\item{x}{Design matrix of x where 1 is to predict the test set, 0 is to be modelled as the train set}

\item{y}{Design matrix of y where 1 is to predict the test set, 0 is to be modelled as the train set}
}
\value{
Prediction framework for gcproc
}
\description{
Can either impute or predict by replacing missing values
}
