% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_config.R
\name{extract_regularise_framework}
\alias{extract_regularise_framework}
\title{Extract regularise framework to put into gcproc}
\usage{
extract_regularise_framework(verbose = T)
}
\value{
Regularisation framework for gcproc
}
\description{
Regularise data to improve modelling interpretability
}
