% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_config.R
\name{extract_transfer_framework}
\alias{extract_transfer_framework}
\title{Extract anchor framework to put into gcproc}
\usage{
extract_transfer_framework(verbose = T)
}
\arguments{
\item{code}{Anchor the code}
}
\value{
Anchor framework for gcproc
}
\description{
Transfers learned parameters from a pre-trained model.
NOTE: This is an empty framework that the user must fill in.
}
