% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcproc.R
\name{gcproc}
\alias{gcproc}
\title{Generalised Canonical Procrustes}
\usage{
gcproc(
  data_list,
  config = gcproc::extract_config(verbose = F),
  transfer = gcproc::extract_transfer_framework(verbose = F),
  recover = gcproc::extract_recovery_framework(verbose = F),
  join = gcproc::extract_join_framework(verbose = F)
)
}
\arguments{
\item{data_list}{List of data matrices of varying dimensionality. Attempts to find similarities among all datasets with a core structure.}

\item{config}{Configuration parameters (required, default provided)}

\item{transfer}{Transferring pre-trained model parameters (not required)}

\item{recover}{Important information used for prediction or imputation (not required)}

\item{join}{Join parameters that share the same axes to be similar (not required)}
}
\value{
Main parameters contains the learned model parameters. The alpha and beta matrix multiply example datasets x and y by, (K)(Y)(v) and (L)(X)(u). By multiplying with the parameter, the dimension of the samples and features can be dimensionally reduced for further visualisation analysis such as embedding or projection.

Code contains the learned shared encoding space. The encoded space refers to the full dimension reduction of both samples and features after matrix multiplication by parameters K and v for y, as well as, L and u for x. The decode is an estimation of the full matrix dataset, where the code is used and matrix multiplied as t(K)(Y_code)t(v), and t(L)(X_code)t(u) to calculate the decoded estimation.

Recover contains the list of predictions for the test dataset as indicated by a 1 in the binary prediction matrices. The prediction occurs in the shared lower dimensional space where all data sets in the list are projected to using a common latent code.
}
\description{
A method that uses a likelihood model to align multiple datasets via an encoding in a lower dimensional space. The parameters can be used to reduce either the feature or the sample dimensions into a smaller subspace for further embedding or prediction. To run as default, only a data list is required - please review the config parameters at gcproc::extract_config(T)  .
}
