% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.gcproc.R
\name{cv.gcproc}
\alias{cv.gcproc}
\title{Run cross-validation on gcproc for regularisation}
\usage{
cv.gcproc(
  x,
  y,
  reference = "y",
  fixed = list(i_dim = F, j_dim = F),
  config = gcproc::extract_config(verbose = F),
  anchors = gcproc::extract_anchors_framework(verbose = F),
  pivots = gcproc::extract_pivots_framework(verbose = F),
  recover = gcproc::extract_recovery_framework(verbose = F)
)
}
\arguments{
\item{x}{Matrix of dataset x}

\item{y}{Matrix of dataset y}

\item{fixed}{Fixed parameters from gcproc}

\item{config}{Configuration parameters from gcproc}

\item{anchors}{Transferring pre-trained model parameters (not required)}

\item{pivots}{Initialisation of model parameters (not required)}

\item{recover}{Recover list from gcproc}
}
\value{
Recovered data from imputation or prediction, with the design matrices and any user input parameters and functions
}
\description{
Evaluate regularisation parameters, by running a search of the tuning parameter space
}
