% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_config.R
\name{extract_fixed_framework}
\alias{extract_fixed_framework}
\title{Extract fixed framework to put into gcproc}
\usage{
extract_fixed_framework(verbose = T)
}
\description{
Fix data to improve modelling capacity for similar axes
}
