% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_config.R
\name{extract_recovery_framework}
\alias{extract_recovery_framework}
\title{Extract recovery framework to put into gcproc}
\usage{
extract_recovery_framework(verbose = T)
}
\arguments{
\item{x}{Design matrix of x where 1 is to predict the test set, 0 is to be modelled as the train set}

\item{y}{Design matrix of y where 1 is to predict the test set, 0 is to be modelled as the train set}

\item{fn}{Allows user to specify a prediction function}

\item{param}{Parameters to be put into prediction function}
}
\value{
Prediction framework for gcproc
}
\description{
Can recover data points by imputing or predicting missing values
}
