% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcproc.R
\name{gcproc}
\alias{gcproc}
\title{Generalised Canonical Procrustes}
\usage{
gcproc(
  x,
  y,
  config = list(k_dim = 70, j_dim = 70, eta = 0.005, max_iter = 1500, min_iter = 15,
    tol = 0.001, log = F, center = T, scale.z = T, batches = 16, cores = 2, verbose = T,
    init = "svd"),
  seed = 1,
  anchors = NULL,
  pivots = NULL
)
}
\arguments{
\item{x}{Reference dataset of cell by gene matrix}

\item{y}{Experimental dataset of cell by gene matrix}

\item{config}{Configuration parameters (please refer to the function extract_config() for more details on each parameter)}

\item{seed}{Fixed seed for random seeding}

\item{anchors}{Fixing and anchoring the main model parameters to transfer prior information}

\item{pivots}{Initialisation of the main model parameters to speed up learning process}
}
\value{
Main parameters contains the learned model parameters.

alpha.L.K:    For dataset y to dimensionally reduce rows to k_dim

alpha.L.J:    For dataset x to dimensionally reduce rows to k_dim

v.beta:       For dataset y to dimensionally reduce rows to j_dim

u.beta:       For dataset x to dimensionally reduce rows to j_dim
}
\description{
A method that uses a likelihood model to align two datasets via an encoding in a lower dimensional space
}
